Fix #49: don't edit concluded experiments in load-experiment
This adds a new storage key which includes all the concluded
experiments and a check in 'load-experiment' which uses that data
to ensure that no changes are made to concluded experiments.

We use the existing '--skip-launched' flag since the intent is
the same in both cases (and concluded experiments have technically
been launched at some point).

We assume that transactions are atomic to ensure consistency
between the existing 'active-experiments' and the new
'concluded-experiments' keys.